⚠️  Update gatsby-source-ghost@4.0.0 ⚠️
This commit updates the dependencies for this starter.

One of the dependencies is gatsby-source-ghost. The update to v4 of this dependency results in the default usage of Ghost Content API v3 as default.

Please read the README to see which setting you need to do to make this starter work with Ghost installations on a lower version than 3.0